add modified tests
