+++ .github/workflows/python.yml
@@ -48,7 +48,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.9]
+        python-version: [3.8]
         os: [ubuntu-latest]
         java: ['11']
         javadist: ['adopt']
