+++ src/main/python/systemds/scuro/aligner/task.py
@@ -21,10 +21,11 @@
 from typing import List
 
 from systemds.scuro.models.model import Model
-
+import numpy as np
+from sklearn.model_selection import KFold
 
 class Task:
-    def __init__(self, name: str, model: Model, labels, train_indices: List, val_indices: List):
+    def __init__(self, name: str, model: Model, labels, train_indices: List, val_indices: List, kfold=5):
         """
         Parent class for the prediction task that is performed on top of the aligned representation
         :param name: Name of the task
@@ -32,12 +33,15 @@ class Task:
         :param labels: Labels used for prediction
         :param train_indices: Indices to extract training data
         :param val_indices: Indices to extract validation data
+        :param kfold: Number of crossvalidation runs
         """
         self.name = name
         self.model = model
         self.labels = labels
         self.train_indices = train_indices
         self.val_indices = val_indices
+        self.kfold = kfold
+        self.counter = 0
 
     def get_train_test_split(self, data):
         X_train = [data[i] for i in self.train_indices]
@@ -49,10 +53,29 @@ class Task:
     
     def run(self, data):
         """
-        The run method need to be implemented by every task class
+        The run method needs to be implemented by every task class
         It handles the training and validation procedures for the specific task
         :param data: The aligned data used in the prediction process
         :return: the validation accuracy
         """
-        pass
+        self.counter += 1
+        skf = KFold(n_splits=self.kfold, shuffle=True, random_state=11)
+        train_scores = []
+        test_scores = []
+        fold = 0
+        X, y, X_test, y_test = self.get_train_test_split(data)
+
+        for train, test in skf.split(X, y):
+            train_X = np.array(X)[train]
+            train_y = np.array(y)[train]
+
+            train_score = self.model.fit(train_X, train_y, X_test, y_test)
+            train_scores.append(train_score)
+
+            test_score = self.model.test(X_test, y_test)
+            test_scores.append(test_score)
+
+            fold += 1
+
+        return [np.mean(train_scores), np.mean(test_scores)]
