+++ src/main/python/systemds/scuro/representations/bert.py
@@ -38,10 +38,10 @@ def read_text_file(file_path):
 class Bert(UnimodalRepresentation):
     def __init__(self, avg_layers=None, output_file=None):
         super().__init__('Bert')
-        
+
         self.avg_layers = avg_layers
         self.output_file = output_file
-    
+
     def parse_all(self, filepath, indices, get_sequences=False):
         # Assumes text is stored in .txt files
         data = []
@@ -54,43 +54,42 @@ class Bert(UnimodalRepresentation):
         else:
             with open(filepath, 'r') as file:
                 data = file.readlines()
-        
+
         model_name = 'bert-base-uncased'
-        tokenizer = BertTokenizer.from_pretrained(model_name)
-        
+        tokenizer = BertTokenizer.from_pretrained(model_name, clean_up_tokenization_spaces=True)
+
         if self.avg_layers is not None:
             model = BertModel.from_pretrained(model_name, output_hidden_states=True)
         else:
             model = BertModel.from_pretrained(model_name)
-        
+
         embeddings = self.create_embeddings(data, model, tokenizer)
-        
+
         if self.output_file is not None:
             data = {}
             for i in range(0, embeddings.shape[0]):
                 data[indices[i]] = embeddings[i]
             self.save_embeddings(data)
-        
+
         return embeddings
-    
+
     def create_embeddings(self, data, model, tokenizer):
         embeddings = []
         for d in data:
             inputs = tokenizer(d, return_tensors="pt", padding=True, truncation=True)
-            
+
             with torch.no_grad():
                 outputs = model(**inputs)
-            
+
             if self.avg_layers is not None:
                 cls_embedding = [outputs.hidden_states[i][:, 0, :] for i in range(-self.avg_layers, 0)]
                 cls_embedding = torch.mean(torch.stack(cls_embedding), dim=0)
             else:
                 cls_embedding = outputs.last_hidden_state[:, 0, :].squeeze().numpy()
-            embeddings.append(cls_embedding)
-        
+            embeddings.append(cls_embedding.numpy())
         embeddings = np.array(embeddings)
         return embeddings.reshape((embeddings.shape[0], embeddings.shape[-1]))
-    
+
     def save_embeddings(self, data):
         with open(self.output_file, 'wb') as file:
             pickle.dump(data, file)
